+++ src/particles/tsnePoint2D.h
@@ -6,16 +6,16 @@
 class TsnePoint2D
 {
 public:
-    glm::vec2 position{0.0f};
-    glm::vec2 derivative{0.0f};
+    glm::dvec2 position{0.0};
+    glm::dvec2 derivative{0.0};
     int label{0};
     #ifdef INDEX_TRACKER
     int ID{0};
     std::uint32_t morton_code{0};
     #endif
 
-    TsnePoint2D(glm::vec2 initPosition, 
-        glm::vec2 initDerivative, 
+    TsnePoint2D(glm::dvec2 initPosition, 
+        glm::dvec2 initDerivative, 
         int initLabel
         #ifdef INDEX_TRACKER
         ,
